Result is bad...
The detect is not so exactly, so the tracking would lose the object after a while.
The color feature can't track all the objects. As to the object cross the screen, the tracking alg can't track it.
The right side of image is a big noise to the tracking alg.