Now the blind zone feature are almost completed.
we keep a list of four points both in configmanager and tracking filter to store the blind zones. And we construct a mask related to each blind zone. The mask is 0 in the blind zone and 1 otherwise. We multiply the mask with frame we got and transform it to the tracking algorithm to implement the blind zone feature.